[change] Activityの起動順序修正
